added base.h, and terrain information with accessors. Tiles now accept this new terrain info. Terrain info will be populated at some point with some sort of db data

+++ hex/InputHandler.cpp
@@ -124,7 +124,7 @@ namespace mandr {
 		mouse_button_released_callback_func = callback;
 	}
 
-	float InputHandler::getMouseDraggedTotalDistance() const {
+	double InputHandler::getMouseDraggedTotalDistance() const {
 		if (!IsMouseBeingDragged())
 			return 0.0f;
 		return sqrt(pow(m_MouseStartDrag.x - m_NewMousePosition.x, 2) + pow(m_MouseStartDrag.y - m_NewMousePosition.y, 2));
